renamed auth module to credentials
