+++ sdk/openai/inference/src/credentials/azure_key_credential.rs
