+++ sdk/openai/inference/src/credentials/mod.rs
