+++ sdk/openai/inference/src/credentials/openai_key_credential.rs
